+++ Blurred-ios/Blurred-ios-Bridging-Header.h
@@ -1,0 +1,10 @@
+//
+//  Header.h
+//  Blurred-ios
+//
+//  Created by Martin Velev on 8/2/20.
+//  Copyright © 2020 BlurrMC. All rights reserved.
+//
+
+#import "GradientNode.h"
+